--- ethereum/build.gradle
@@ -1,1 +1,0 @@
-jar { enabled = false }
