+++ ethereum/executionlayer/src/main/java/tech/pegasys/teku/ethereum/executionlayer/ExecutionEngineChannelImpl.java
@@ -1,0 +1,146 @@
+/*
+ * Copyright 2021 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.ethereum.executionlayer;
+
+import static com.google.common.base.Preconditions.checkArgument;
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import java.util.Optional;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.ethereum.executionlayer.client.ExecutionEngineClient;
+import tech.pegasys.teku.ethereum.executionlayer.client.Web3JExecutionEngineClient;
+import tech.pegasys.teku.ethereum.executionlayer.client.schema.ExecutionPayloadV1;
+import tech.pegasys.teku.ethereum.executionlayer.client.schema.ForkChoiceStateV1;
+import tech.pegasys.teku.ethereum.executionlayer.client.schema.ForkChoiceUpdatedResult;
+import tech.pegasys.teku.ethereum.executionlayer.client.schema.PayloadAttributesV1;
+import tech.pegasys.teku.ethereum.executionlayer.client.schema.Response;
+import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.spec.Spec;
+import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
+import tech.pegasys.teku.spec.datastructures.execution.PowBlock;
+import tech.pegasys.teku.spec.executionengine.ExecutePayloadResult;
+import tech.pegasys.teku.spec.executionengine.ExecutionEngineChannel;
+import tech.pegasys.teku.spec.executionengine.ForkChoiceState;
+import tech.pegasys.teku.spec.executionengine.PayloadAttributes;
+import tech.pegasys.teku.spec.schemas.SchemaDefinitionsMerge;
+import tech.pegasys.teku.ssz.type.Bytes8;
+
+public class ExecutionEngineChannelImpl implements ExecutionEngineChannel {
+  private static final Logger LOG = LogManager.getLogger();
+
+  private final ExecutionEngineClient executionEngineClient;
+  private final Spec spec;
+
+  public static ExecutionEngineChannelImpl create(String eeEndpoint, Spec spec) {
+    checkNotNull(eeEndpoint);
+    return new ExecutionEngineChannelImpl(new Web3JExecutionEngineClient(eeEndpoint), spec);
+  }
+
+  public ExecutionEngineChannelImpl(ExecutionEngineClient executionEngineClient, Spec spec) {
+    this.spec = spec;
+    this.executionEngineClient = executionEngineClient;
+  }
+
+  private static <K> K unwrapResponseOrThrow(Response<K> response) {
+    checkArgument(
+        response.getErrorMessage() == null,
+        "Invalid remote response: %s",
+        response.getErrorMessage());
+    return response.getPayload();
+  }
+
+  @Override
+  public SafeFuture<Optional<PowBlock>> getPowBlock(final Bytes32 blockHash) {
+    LOG.trace("calling getPowBlock(blockHash={})", blockHash);
+
+    return executionEngineClient
+        .getPowBlock(blockHash)
+        .thenPeek(powBlock -> LOG.trace("getPowBlock(blockHash={}) -> {}", blockHash, powBlock));
+  }
+
+  @Override
+  public SafeFuture<PowBlock> getPowChainHead() {
+    LOG.trace("calling getPowChainHead()");
+
+    return executionEngineClient
+        .getPowChainHead()
+        .thenPeek(powBlock -> LOG.trace("getPowChainHead() -> {}", powBlock));
+  }
+
+  @Override
+  public SafeFuture<tech.pegasys.teku.spec.executionengine.ForkChoiceUpdatedResult>
+      forkChoiceUpdated(
+          final ForkChoiceState forkChoiceState,
+          final Optional<PayloadAttributes> payloadAttributes) {
+
+    LOG.trace(
+        "calling forkChoiceUpdated(forkChoiceState={}, payloadAttributes={})",
+        forkChoiceState,
+        payloadAttributes);
+
+    return executionEngineClient
+        .forkChoiceUpdated(
+            ForkChoiceStateV1.fromInternalForkChoiceState(forkChoiceState),
+            PayloadAttributesV1.fromInternalForkChoiceState(payloadAttributes))
+        .thenApply(ExecutionEngineChannelImpl::unwrapResponseOrThrow)
+        .thenApply(ForkChoiceUpdatedResult::asInternalExecutionPayload)
+        .thenPeek(
+            forkChoiceUpdatedResult ->
+                LOG.trace(
+                    "forkChoiceUpdated(forkChoiceState={}, payloadAttributes={}) -> {}",
+                    forkChoiceState,
+                    payloadAttributes,
+                    forkChoiceUpdatedResult));
+  }
+
+  @Override
+  public SafeFuture<ExecutionPayload> getPayload(final Bytes8 payloadId, final UInt64 slot) {
+    LOG.trace("calling getPayload(payloadId={}, slot={})", payloadId, slot);
+
+    return executionEngineClient
+        .getPayload(payloadId)
+        .thenApply(ExecutionEngineChannelImpl::unwrapResponseOrThrow)
+        .thenCombine(
+            SafeFuture.of(
+                () ->
+                    SchemaDefinitionsMerge.required(spec.atSlot(slot).getSchemaDefinitions())
+                        .getExecutionPayloadSchema()),
+            ExecutionPayloadV1::asInternalExecutionPayload)
+        .thenPeek(
+            executionPayload ->
+                LOG.trace(
+                    "getPayload(payloadId={}, slot={}) -> {}", payloadId, slot, executionPayload));
+  }
+
+  @Override
+  public SafeFuture<ExecutePayloadResult> executePayload(final ExecutionPayload executionPayload) {
+    LOG.trace("calling executePayload(executionPayload={})", executionPayload);
+
+    return executionEngineClient
+        .executePayload(ExecutionPayloadV1.fromInternalExecutionPayload(executionPayload))
+        .thenApply(ExecutionEngineChannelImpl::unwrapResponseOrThrow)
+        .thenApply(
+            tech.pegasys.teku.ethereum.executionlayer.client.schema.ExecutePayloadResult
+                ::asInternalExecutionPayload)
+        .thenPeek(
+            executePayloadResult ->
+                LOG.trace(
+                    "executePayload(executionPayload={}) -> {}",
+                    executionPayload,
+                    executePayloadResult));
+  }
+}
