+++ ethereum/executionlayer/src/main/java/tech/pegasys/teku/ethereum/executionlayer/client/schema/ExecutionPayloadV1.java
@@ -1,0 +1,247 @@
+/*
+ * Copyright 2021 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.ethereum.executionlayer.client.schema;
+
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.google.common.base.MoreObjects;
+import java.util.List;
+import java.util.Objects;
+import java.util.stream.Collectors;
+import org.apache.tuweni.bytes.Bytes;
+import org.apache.tuweni.bytes.Bytes32;
+import org.apache.tuweni.units.bigints.UInt256;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.Bytes20Deserializer;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.Bytes20Serializer;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.Bytes32Deserializer;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.BytesDeserializer;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.BytesSerializer;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.UInt256AsHexDeserializer;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.UInt256AsHexSerializer;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.UInt64AsHexDeserializer;
+import tech.pegasys.teku.ethereum.executionlayer.client.serialization.UInt64AsHexSerializer;
+import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
+import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadSchema;
+import tech.pegasys.teku.ssz.collections.impl.SszByteListImpl;
+import tech.pegasys.teku.ssz.type.Bytes20;
+
+public class ExecutionPayloadV1 {
+
+  @JsonSerialize(using = BytesSerializer.class)
+  @JsonDeserialize(using = Bytes32Deserializer.class)
+  public final Bytes32 parentHash;
+
+  @JsonSerialize(using = Bytes20Serializer.class)
+  @JsonDeserialize(using = Bytes20Deserializer.class)
+  public final Bytes20 coinbase;
+
+  @JsonSerialize(using = BytesSerializer.class)
+  @JsonDeserialize(using = Bytes32Deserializer.class)
+  public final Bytes32 stateRoot;
+
+  @JsonSerialize(using = BytesSerializer.class)
+  @JsonDeserialize(using = Bytes32Deserializer.class)
+  public final Bytes32 receiptRoot;
+
+  @JsonSerialize(using = BytesSerializer.class)
+  @JsonDeserialize(using = BytesDeserializer.class)
+  public final Bytes logsBloom;
+
+  @JsonSerialize(using = BytesSerializer.class)
+  @JsonDeserialize(using = Bytes32Deserializer.class)
+  public final Bytes32 random;
+
+  @JsonSerialize(using = UInt64AsHexSerializer.class)
+  @JsonDeserialize(using = UInt64AsHexDeserializer.class)
+  public final UInt64 blockNumber;
+
+  @JsonSerialize(using = UInt64AsHexSerializer.class)
+  @JsonDeserialize(using = UInt64AsHexDeserializer.class)
+  public final UInt64 gasLimit;
+
+  @JsonSerialize(using = UInt64AsHexSerializer.class)
+  @JsonDeserialize(using = UInt64AsHexDeserializer.class)
+  public final UInt64 gasUsed;
+
+  @JsonSerialize(using = UInt64AsHexSerializer.class)
+  @JsonDeserialize(using = UInt64AsHexDeserializer.class)
+  public final UInt64 timestamp;
+
+  @JsonSerialize(using = BytesSerializer.class)
+  @JsonDeserialize(using = BytesDeserializer.class)
+  public final Bytes extraData;
+
+  @JsonSerialize(using = UInt256AsHexSerializer.class)
+  @JsonDeserialize(using = UInt256AsHexDeserializer.class)
+  public final UInt256 baseFeePerGas;
+
+  @JsonSerialize(using = BytesSerializer.class)
+  @JsonDeserialize(using = Bytes32Deserializer.class)
+  public final Bytes32 blockHash;
+
+  @JsonSerialize(contentUsing = BytesSerializer.class)
+  @JsonDeserialize(contentUsing = BytesDeserializer.class)
+  public final List<Bytes> transactions;
+
+  public ExecutionPayloadV1(
+      @JsonProperty("parentHash") Bytes32 parentHash,
+      @JsonProperty("coinbase") Bytes20 coinbase,
+      @JsonProperty("stateRoot") Bytes32 stateRoot,
+      @JsonProperty("receiptRoot") Bytes32 receiptRoot,
+      @JsonProperty("logsBloom") Bytes logsBloom,
+      @JsonProperty("random") Bytes32 random,
+      @JsonProperty("blockNumber") UInt64 blockNumber,
+      @JsonProperty("gasLimit") UInt64 gasLimit,
+      @JsonProperty("gasUsed") UInt64 gasUsed,
+      @JsonProperty("timestamp") UInt64 timestamp,
+      @JsonProperty("extraData") Bytes extraData,
+      @JsonProperty("baseFeePerGas") UInt256 baseFeePerGas,
+      @JsonProperty("blockHash") Bytes32 blockHash,
+      @JsonProperty("transactions") List<Bytes> transactions) {
+    checkNotNull(parentHash, "parentHash");
+    checkNotNull(coinbase, "coinbase");
+    checkNotNull(stateRoot, "stateRoot");
+    checkNotNull(receiptRoot, "receiptRoot");
+    checkNotNull(logsBloom, "logsBloom");
+    checkNotNull(random, "random");
+    checkNotNull(blockNumber, "blockNumber");
+    checkNotNull(gasLimit, "gasLimit");
+    checkNotNull(gasUsed, "gasUsed");
+    checkNotNull(timestamp, "timestamp");
+    checkNotNull(extraData, "extraData");
+    checkNotNull(baseFeePerGas, "baseFeePerGas");
+    checkNotNull(blockHash, "blockHash");
+    this.parentHash = parentHash;
+    this.coinbase = coinbase;
+    this.stateRoot = stateRoot;
+    this.receiptRoot = receiptRoot;
+    this.logsBloom = logsBloom;
+    this.random = random;
+    this.blockNumber = blockNumber;
+    this.gasLimit = gasLimit;
+    this.gasUsed = gasUsed;
+    this.timestamp = timestamp;
+    this.extraData = extraData;
+    this.baseFeePerGas = baseFeePerGas;
+    this.blockHash = blockHash;
+    this.transactions = transactions != null ? transactions : List.of();
+  }
+
+  public ExecutionPayload asInternalExecutionPayload(
+      ExecutionPayloadSchema executionPayloadSchema) {
+    return executionPayloadSchema.create(
+        parentHash,
+        coinbase,
+        stateRoot,
+        receiptRoot,
+        logsBloom,
+        random,
+        blockNumber,
+        gasLimit,
+        gasUsed,
+        timestamp,
+        extraData,
+        baseFeePerGas,
+        blockHash,
+        transactions);
+  }
+
+  public static ExecutionPayloadV1 fromInternalExecutionPayload(ExecutionPayload executionPayload) {
+    return new ExecutionPayloadV1(
+        executionPayload.getParentHash(),
+        executionPayload.getCoinbase(),
+        executionPayload.getStateRoot(),
+        executionPayload.getReceiptRoot(),
+        executionPayload.getLogsBloom(),
+        executionPayload.getRandom(),
+        executionPayload.getBlockNumber(),
+        executionPayload.getGasLimit(),
+        executionPayload.getGasUsed(),
+        executionPayload.getTimestamp(),
+        executionPayload.getExtraData(),
+        executionPayload.getBaseFeePerGas(),
+        executionPayload.getBlockHash(),
+        executionPayload.getTransactions().stream()
+            .map(SszByteListImpl::getBytes)
+            .collect(Collectors.toList()));
+  }
+
+  @Override
+  public boolean equals(final Object o) {
+    if (this == o) {
+      return true;
+    }
+    if (o == null || getClass() != o.getClass()) {
+      return false;
+    }
+    final ExecutionPayloadV1 that = (ExecutionPayloadV1) o;
+    return Objects.equals(parentHash, that.parentHash)
+        && Objects.equals(coinbase, that.coinbase)
+        && Objects.equals(stateRoot, that.stateRoot)
+        && Objects.equals(receiptRoot, that.receiptRoot)
+        && Objects.equals(logsBloom, that.logsBloom)
+        && Objects.equals(random, that.random)
+        && Objects.equals(blockNumber, that.blockNumber)
+        && Objects.equals(gasLimit, that.gasLimit)
+        && Objects.equals(gasUsed, that.gasUsed)
+        && Objects.equals(timestamp, that.timestamp)
+        && Objects.equals(extraData, that.extraData)
+        && Objects.equals(baseFeePerGas, that.baseFeePerGas)
+        && Objects.equals(blockHash, that.blockHash)
+        && Objects.equals(transactions, that.transactions);
+  }
+
+  @Override
+  public int hashCode() {
+    return Objects.hash(
+        parentHash,
+        coinbase,
+        stateRoot,
+        receiptRoot,
+        logsBloom,
+        random,
+        blockNumber,
+        gasLimit,
+        gasUsed,
+        timestamp,
+        extraData,
+        baseFeePerGas,
+        blockHash,
+        transactions);
+  }
+
+  @Override
+  public String toString() {
+    return MoreObjects.toStringHelper(this)
+        .add("parentHash", parentHash)
+        .add("coinbase", coinbase)
+        .add("stateRoot", stateRoot)
+        .add("receiptRoot", receiptRoot)
+        .add("logsBloom", logsBloom)
+        .add("random", random)
+        .add("blockNumber", blockNumber)
+        .add("gasLimit", gasLimit)
+        .add("gasUsed", gasUsed)
+        .add("timestamp", timestamp)
+        .add("extraData", extraData)
+        .add("baseFeePerGas", baseFeePerGas)
+        .add("blockHash", blockHash)
+        .add("transactions", transactions)
+        .toString();
+  }
+}
