--- services/powchain/src/main/java/tech/pegasys/teku/services/executionengine/client/serialization/Bytes20Serializer.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2021 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.services.executionengine.client.serialization;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import java.io.IOException;
-import tech.pegasys.teku.ssz.type.Bytes20;
-
-public class Bytes20Serializer extends JsonSerializer<Bytes20> {
-  @Override
-  public void serialize(Bytes20 value, JsonGenerator gen, SerializerProvider serializers)
-      throws IOException {
-    gen.writeString(value.toHexString().toLowerCase());
-  }
-}
