--- services/powchain/src/main/java/tech/pegasys/teku/services/executionengine/client/serialization/Bytes8Deserializer.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2021 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.services.executionengine.client.serialization;
-
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import java.io.IOException;
-import tech.pegasys.teku.ssz.type.Bytes8;
-
-public class Bytes8Deserializer extends JsonDeserializer<Bytes8> {
-
-  @Override
-  public Bytes8 deserialize(JsonParser p, DeserializationContext ctxt) throws IOException {
-    return Bytes8.fromHexString(p.getValueAsString());
-  }
-}
