--- services/powchain/src/main/java/tech/pegasys/teku/services/executionengine/client/serialization/UInt64AsHexSerializer.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2021 ConsenSys AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
- * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package tech.pegasys.teku.services.executionengine.client.serialization;
-
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
-import java.io.IOException;
-import org.apache.tuweni.bytes.Bytes;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-
-public class UInt64AsHexSerializer extends JsonSerializer<UInt64> {
-  @Override
-  public void serialize(UInt64 value, JsonGenerator gen, SerializerProvider serializers)
-      throws IOException {
-    gen.writeString(toHexString(value));
-  }
-
-  public static String toHexString(UInt64 value) {
-    return Bytes.ofUnsignedLong(value.longValue()).toQuantityHexString();
-  }
-}
