+++ settings.gradle
@@ -6,6 +6,7 @@ include 'ethereum:core'
 include 'ethereum:spec'
 include 'ethereum:dataproviders'
 include 'ethereum:events'
+include 'ethereum:executionlayer'
 include 'ethereum:networks'
 include 'ethereum:pow:api'
 include 'ethereum:statetransition'
@@ -38,6 +39,7 @@ include 'networking:eth2'
 include 'networking:nat'
 include 'networking:p2p'
 include 'services'
+include 'services:executionengine'
 include 'services:powchain'
 include 'services:beaconchain'
 include 'services:chainstorage'
