+++ teku/build.gradle
@@ -30,6 +30,7 @@ dependencies {
   implementation project(':pow')
   implementation project(':services:beaconchain')
   implementation project(':services:chainstorage')
+  implementation project(':services:executionengine')
   implementation project(':services:powchain')
   implementation project(':infrastructure:serviceutils')
   implementation project(':services:timer')
