+++ teku/src/main/java/tech/pegasys/teku/config/TekuConfiguration.java
@@ -93,8 +93,6 @@ public class TekuConfiguration {
             syncConfig,
             beaconRestApiConfig,
             powchainConfiguration,
-            executionEngineConfiguration,
-            loggingConfig,
             storeConfig,
             spec);
     this.validatorClientConfig =
