+++ teku/src/main/java/tech/pegasys/teku/services/BeaconNodeServiceController.java
@@ -19,6 +19,7 @@ import tech.pegasys.teku.networking.nat.NatService;
 import tech.pegasys.teku.service.serviceutils.ServiceConfig;
 import tech.pegasys.teku.services.beaconchain.BeaconChainService;
 import tech.pegasys.teku.services.chainstorage.StorageService;
+import tech.pegasys.teku.services.executionengine.ExecutionEngineService;
 import tech.pegasys.teku.services.powchain.PowchainService;
 import tech.pegasys.teku.services.timer.TimerService;
 import tech.pegasys.teku.validator.client.ValidatorClientService;
@@ -37,6 +38,9 @@ public class BeaconNodeServiceController extends ServiceController {
             tekuConfig.natConfiguration(),
             tekuConfig.network().getListenPort(),
             tekuConfig.discovery().isDiscoveryEnabled()));
+    if (tekuConfig.executionEngine().isEnabled()) {
+      services.add(new ExecutionEngineService(serviceConfig, tekuConfig.executionEngine()));
+    }
     powchainService(tekuConfig, serviceConfig).ifPresent(services::add);
   }
