+++ teku/src/test/java/tech/pegasys/teku/cli/BeaconNodeCommandTest.java
@@ -450,7 +450,7 @@ public class BeaconNodeCommandTest extends AbstractBeaconNodeCommandTest {
 
     return expectedConfigurationBuilder()
         .eth2NetworkConfig(b -> b.applyNetworkDefaults("mainnet"))
-        .executionEngine(b -> b.endpoints(new ArrayList<String>()))
+        .executionEngine(b -> b.endpoint(null))
         .powchain(
             b -> {
               b.depositContract(networkConfig.getEth1DepositContractAddress());
@@ -500,7 +500,7 @@ public class BeaconNodeCommandTest extends AbstractBeaconNodeCommandTest {
   private TekuConfiguration.Builder expectedConfigurationBuilder() {
     return TekuConfiguration.builder()
         .eth2NetworkConfig(b -> b.applyMinimalNetworkDefaults().eth1DepositContractAddress(address))
-        .executionEngine(b -> b.endpoints(List.of("http://localhost:8550")))
+        .executionEngine(b -> b.endpoint("http://localhost:8550"))
         .powchain(
             b ->
                 b.eth1Endpoints(List.of("http://localhost:8545"))
